Fixed Auth setup, Fixed backend to actually be called and working

+++ frontend/src/hooks/useAuth.ts
@@ -1,12 +1,9 @@
 // src/hooks/useAuth.ts
-import { useState } from 'react'
 import { createUserWithEmailAndPassword, signInWithEmailAndPassword, updateProfile, UserCredential } from 'firebase/auth';
 import { FIREBASE_AUTH } from '@/src/config/FirebaseConfig';
 import { setUserInformation } from '@/src/utils/auth';
 
 export const useAuth = () => {
-  const [error, setError] = useState('')
-  
   const handleAuth = async (email: string, password: string, username: string, isSignUp: boolean) => {
     try {
       let userCredential: UserCredential | null = null;
@@ -26,16 +23,13 @@ export const useAuth = () => {
       } catch (error) {
         // Handle errors for both sign-in and sign-up
         if (!isSignUp && error.code === 'auth/invalid-credential') {
-          setError('Invalid email or password. Please try again.');
           return { success: false, error: 'Invalid email or password. Please try again.' };
         }
       
         if (isSignUp && error.code === 'auth/email-already-in-use') {
-          setError('Email already in use. Please use a different email.');
           return { success: false, error: 'Email already in use. Please use a different email.' };
         }
-      
-        setError('Authentication error: ' + error.message);
+
         return { success: false, error: 'Authentication error: ' + error.message };
       }
 
@@ -43,7 +37,7 @@ export const useAuth = () => {
       const idToken = await userCredential.user.getIdToken();
 
       // Send the ID token and username to the backend for registration/login
-      const response = await fetch(`${process.env.BACKEND_URL}/login`, {
+      const response = await fetch(`${process.env.EXPO_PUBLIC_BACKEND_URL}/login`, {
         method: 'POST',
         headers: {
           'Authorization': `Bearer ${idToken}`,
@@ -52,23 +46,19 @@ export const useAuth = () => {
         body: JSON.stringify({ username }),
       });
 
-      console.log(response)
-
       if (!response.ok) {
         // Handle failure and delete the newly created user if registration fails
-        if (userCredential && !userCredential.user.emailVerified) {
+        if (isSignUp && userCredential && !userCredential.user.emailVerified) {
           await userCredential.user.delete();
         }
-        setError('Authentication failed. Please try again.');
         return { success: false, error: 'Authentication failed. Please try again.' };
       }
 
       return { success: true };
     } catch (error) {
-      setError('Authentication failed. Please try again.');
-      return { success: false, error: error.message };
+      return { success: false, error: 'Authentication failed. Please try again.' };
     }
   }
 
-  return { handleAuth, error };
+  return { handleAuth };
 }
